// apps/web/src/api/generated/HfSearchResultDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BackendCompatMapDto } from "./BackendCompatMapDto";
import type { RepoFileDto } from "./RepoFileDto";

export type HfSearchResultDto = { repo_id: string, author: string | null, license: string | null, downloads_30d: bigint | null, last_modified: string | null, files: Array<RepoFileDto>, formats: Array<string>, quantizations: Array<string>, backend_compat: BackendCompatMapDto, already_installed_ids: Array<string>, };